Move workspace package to internal/runner/workspace
The workspace package is primarily used by the runner and its
subpackages. Move it under internal/runner/ to better reflect
this relationship and improve package organization.
