+++ internal/command/setup.go
@@ -10,7 +10,7 @@ import (
 	"github.com/icholy/xagent/internal/configfile"
 	"github.com/icholy/xagent/internal/auth/deviceauth"
 	xagentv1 "github.com/icholy/xagent/internal/proto/xagent/v1"
-	"github.com/icholy/xagent/internal/workspace"
+	"github.com/icholy/xagent/internal/runner/workspace"
 	"github.com/icholy/xagent/internal/xagentclient"
 	"github.com/urfave/cli/v3"
 	"github.com/zitadel/oidc/v3/pkg/oidc"
